requirements.txt
newspaper3k
feedparser
pandas
numpy==1.24.4
dateparser
nltk
telethon
lxml_html_clean
openpyxl
pdfplumber 
camelot-py[pdf]
transformers
torch
yfinance
fpdf
vaderSentiment
scikit-learn
beautifulsoup4
requests
xgboost
lightgbm
tensorflow
ta-lib
matplotlib
seaborn
apache-airflow